Format main CSS
For main, set min-height so it can always contain at least the full page of a device. then I used display grid to ensure for most mobile devices except tablets, image and nav are well placed.

The img-hero-img height and width is set to 100%. It is a bit skewwed but looks like what is on Figma. if the team agrees, We can use object-fit:cover to fix the skewness. The grid-row is just to ensure the image position.

Nav is set to width of 100% and color of Dark Grey based on design color.

I ensured the svg was filled with the right color. And also, p, svg and img are all on the same level visually.